COMPATIBLE_MACHINES for debugpy
Without this setting, the build meta-protos for other
architectures throws an error even when debugpy is not build.
This is caused by the MACHINE_ARCH test during parsing for each architecture.
Even with the error, the build is successfull for unsupported
architectures.

The correct way to handle this would be to ignore/not parse this recipe
for non supported architectures. With this change, any build depending
on debugpy would fail, but independend build work as expected.